refactor: BookPreviewApi_SSR과 일반 API 통합
